Project Files/requirements.txt
flask
tensorflow
cloudant
Pillow
numpy
werkzeug
python-dotenv
